react refactor part 2~
