+++ src/components/Wikidiv/index.js
@@ -3,22 +3,15 @@
 const React = require('react')
 require('./main.css')
 
-type WikiProps = {
-  onClick: Function,
-  shown: boolean
-}
-
-const WikiDiv = (props: WikiProps) => {
+const WikiDiv = () => {
   const iframe = React.createElement('iframe', { 
     name: 'wiki', 
     id: 'wikiframe', 
-    src: '',
-    onClick: () => console.log('clicked!')
+    src: ''
   })
- 
-  return React.createElement('div', { 
-    className: props.shown ? '' : 'hidden',
-    onClick: () => console.log('aaaa') //, props.onClick
+
+  return React.createElement('div', {
+    id: 'wikidiv',
   }, iframe)
 }
