+++ src/components/Wikidiv/main.css
@@ -1,13 +1,41 @@
 #wikidiv {
   width: 35%;
-  height: 100%;
+  height: calc(100% - 60px);
 	float: right;
 }
-
 
 #wikidiv > div {
   width: 100%;
   height: 100%;
+}
+
+.collapse {
+  height: 60px;
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+  box-sizing: border-box;
+  padding-right: 20px;
+}
+
+.collapse span {
+  font-size: 2em;
+  display: inline-block;
+  height: 40px;
+  width: 40px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background-color: #54a0ff;
+  border-radius: 50%;
+  color: white;
+  box-shadow: 2px 2px 6px 3px rgba(0, 0, 0, 0.15);
+  transition: background-color 0.2s, box-shadow 0.2s;
+}
+
+.collapse span:hover {
+  background-color: #2e86de;
+  box-shadow: 2px 2px 8px 5px rgba(0, 0, 0, 0.1);
 }
 
 @media only screen and (min-width: 1200px) {
