+++ static/tsomi.css
@@ -204,7 +204,7 @@ path.link {
 }
 
 @media only screen and (min-width: 1500px) {
-  #wikidiv {
-    width: 25%;
+  #chartdiv {
+    width: 75%;
   }
 }
